Move executors to docker(-cloud)-plugin
This moves the executors from the master, over to run in dynamically
created docker containers.

This needs patched docker-jenkins-job image with java added.

This uses the old localhost-remapping technique to have docker
containers access other docker containers exposed ports.

To simplify, that same technique for access to bootstrap-docker-builds
git repo.

This also drops the workarounds for docker-custom-build-environment,
which doesn't work in the docker-cloud based agents, which enviorment
cant be matched to the hosts enviorment to get the plugin working.
This is re-implemented in "Move run-in-container jobs to run on
docker-cloud" in the bootstrap-docker-builds repo.

The docker-plugin uses the docker api directly, so we don't need the
docker binary any longer, just the socket.