Re-introduce to fix docker-custom-build-environment
docker-custom-build-environment blindly volume mounts bits from the
agent into the custom-container it uses, and this doesn't work that well
in a docker in docker setup, where the agents containers paths might not
match what the host uses.
This (re)introduces enough bits to have the host and the master agent
executors match close enough to get docker-custom-build-environment to
work.