Approve the groovy bits our pipeline code needs
In the current iteration of pipeline code, the only groovy method we
need to white list is the toBoolean(), to convert strings to booleans in
a sane matter. All other needs have bin removed by rewriting the
pipeline so it uses pipeline native directives to check out the repo and
read the .jenkins.yaml from there.

Older iterations of the pipeline code will need more white lists, but
because some of them are somewhat dangerous, those have bin squished
away, but they are quite easy to figure out for anyone with a interest
in code archaeology.